+++ cmd/remove.go
@@ -21,7 +21,7 @@ var removeCmd = &cobra.Command{
 			return err
 		}
 
-		u.Dim("crib " + Version)
+		u.Dim(versionString())
 
 		if err := eng.Remove(cmd.Context(), ws); err != nil {
 			return err